@@ -1,34 +1,19 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <IsPackable>false</IsPackable>
-    <TargetFrameworks>netstandard2.0;net5.0;net6.0</TargetFrameworks>
-    <LangVersion>latest</LangVersion>
+<Project Sdk="Microsoft.NET.Sdk">
 
-    <AssemblyName>SwedbankPay.Sdk</AssemblyName>
-    <RootNamespace>SwedbankPay.Sdk</RootNamespace>
-
-    <Authors>Swedbank Pay</Authors>
-    <Description>A client library and SDK for Swedbank Pay's RESTful API Platform</Description>
-    <Copyright>Swedbank Pay</Copyright>
-
-    <PackageId>SwedbankPay.Sdk.NotPacked</PackageId>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-    <DocumentationFile>SwedbankPay.Sdk.xml</DocumentationFile>
-  </PropertyGroup>
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-      <_Parameter1>$(MSBuildProjectName).Tests</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-  <ItemGroup>
-    <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-      <_Parameter1>$(AssemblyName).Infrastructure</_Parameter1>
-    </AssemblyAttribute>
-  </ItemGroup>
-
-  <ItemGroup>
-    <PackageReference Include="System.Text.Json" Version="[5.0.0,7.0.0)" />
-  </ItemGroup>
-
+    <PropertyGroup>
+        <TargetFramework>net7.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+    
+    <ItemGroup>
+        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+            <_Parameter1>$(MSBuildProjectName).Tests</_Parameter1>
+        </AssemblyAttribute>
+    </ItemGroup>
+    <ItemGroup>
+        <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
+            <_Parameter1>$(AssemblyName).Infrastructure</_Parameter1>
+        </AssemblyAttribute>
+    </ItemGroup>
 </Project>
