@@ -1,28 +1,11 @@
-﻿using SwedbankPay.Sdk.Consumers;
-using SwedbankPay.Sdk.PaymentInstruments;
-using SwedbankPay.Sdk.PaymentOrders;
+using SwedbankPay.Sdk.PaymentOrder;
 
-namespace SwedbankPay.Sdk
+namespace SwedbankPay.Sdk;
+
+public interface ISwedbankPayClient
 {
     /// <summary>
-    /// The entrypoint of this SDK!
-    /// Used to access the different APIs'.
+    /// Resource to create and get payment orders.
     /// </summary>
-    public interface ISwedbankPayClient
-    {
-        /// <summary>
-        /// Resource to create and get payment orders.
-        /// </summary>
-        IPaymentOrdersResource PaymentOrders { get; }
-
-        /// <summary>
-        /// Resource to access consumer information.
-        /// </summary>
-        IConsumersResource Consumers { get; }
-
-        /// <summary>
-        /// Resource to create and get payments on several payment instruments.
-        /// </summary>
-        IPaymentInstrumentsResource Payments { get; }
-    }
+    IPaymentOrdersResource PaymentOrders { get; }
 }
