@@ -1,40 +1,38 @@
-﻿using System;
 using System.Globalization;
 
-namespace SwedbankPay.Sdk
+namespace SwedbankPay.Sdk;
+
+/// <summary>
+/// Contains information about a payers prefered langauge.
+/// </summary>
+public class Language
 {
     /// <summary>
-    /// Contains information about a payers prefered langauge.
+    /// Instantiates a new <seealso cref="Language"/>.
     /// </summary>
-    public class Language
+    /// <param name="language">A string in the format of xx-YY</param>
+    public Language(string language)
     {
-        /// <summary>
-        /// Instantiates a new <seealso cref="Language"/>.
-        /// </summary>
-        /// <param name="language">A string in the format of xx-YY</param>
-        public Language(string language)
+        if (language == null)
         {
-            if (language == null)
-            {
-                throw new ArgumentNullException(nameof(language));
-            }
-
-            Culture = new CultureInfo(language);
-            if (Culture.IsNeutralCulture)
-            {
-                throw new ArgumentException($"Must be given in a xx-YY format.", nameof(language));
-            }
+            throw new ArgumentNullException(nameof(language));
         }
 
-        private CultureInfo Culture { get; }
-
-        /// <summary>
-        /// Returns the provided language from the constructor as a string.
-        /// </summary>
-        /// <returns>The provided language from the constructor.</returns>
-        public override string ToString()
+        Culture = new CultureInfo(language);
+        if (Culture.IsNeutralCulture)
         {
-            return Culture.Name;
+            throw new ArgumentException($"Must be given in a xx-YY format.", nameof(language));
         }
+    }
+
+    private CultureInfo Culture { get; }
+
+    /// <summary>
+    /// Returns the provided language from the constructor as a string.
+    /// </summary>
+    /// <returns>The provided language from the constructor.</returns>
+    public override string ToString()
+    {
+        return Culture.Name;
     }
 }
