@@ -1,95 +1,45 @@
-﻿using System.Collections;
-using System.Collections.Generic;
+using System.Collections;
 
-namespace SwedbankPay.Sdk
+namespace SwedbankPay.Sdk;
+
+public class OperationsBase : IDictionary<LinkRelation, HttpOperation>
 {
-    /// <summary>
-    /// Abstract class to handle the relation between <seealso cref="LinkRelation"/> and <seealso cref="HttpOperation"/>.
-    /// </summary>
-    public abstract class OperationsBase : IDictionary<LinkRelation, HttpOperation>
-    {
-        private readonly Dictionary<LinkRelation, HttpOperation> internalDictionary = new Dictionary<LinkRelation, HttpOperation>();
+    private readonly Dictionary<LinkRelation, HttpOperation?> internalDictionary = new();
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public ICollection<LinkRelation> Keys => this.internalDictionary.Keys;
+    public IEnumerator<KeyValuePair<LinkRelation, HttpOperation>> GetEnumerator() => internalDictionary.GetEnumerator();
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public ICollection<HttpOperation> Values => this.internalDictionary.Values;
+    IEnumerator IEnumerable.GetEnumerator() => internalDictionary.GetEnumerator();
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public int Count => this.internalDictionary.Count;
+    public void Add(KeyValuePair<LinkRelation, HttpOperation> item) => internalDictionary[item.Key] = item.Value;
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public bool IsReadOnly => false;
 
-        HttpOperation IDictionary<LinkRelation, HttpOperation>.this[LinkRelation rel] { get => this.internalDictionary.ContainsKey(rel) ? this.internalDictionary[rel] : null; set => this.internalDictionary[rel] = value; }
+    public void Clear() => internalDictionary.Clear();
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public IEnumerator GetEnumerator() => this.internalDictionary.GetEnumerator();
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public void Add(LinkRelation key, HttpOperation value) => this.internalDictionary.Add(key, value);
+    public bool Contains(KeyValuePair<LinkRelation, HttpOperation> item) => internalDictionary.ContainsKey(item.Key) &&
+                                                                            internalDictionary[item.Key]
+                                                                                .Equals(item.Value);
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public bool ContainsKey(LinkRelation key) => this.internalDictionary.ContainsKey(key);
+    public void CopyTo(KeyValuePair<LinkRelation, HttpOperation>[] array, int arrayIndex) =>
+        ((IDictionary)internalDictionary).CopyTo(array, arrayIndex);
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public bool Remove(LinkRelation key) => this.internalDictionary.Remove(key);
+    public bool Remove(KeyValuePair<LinkRelation, HttpOperation> item) => internalDictionary.Remove(item.Key);
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public bool TryGetValue(LinkRelation key, out HttpOperation value) => this.internalDictionary.TryGetValue(key, out value);
+    public int Count => internalDictionary.Count;
+    public bool IsReadOnly => false;
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public void Add(KeyValuePair<LinkRelation, HttpOperation> item) => this.internalDictionary[item.Key] = item.Value;
+    public void Add(LinkRelation key, HttpOperation value) => internalDictionary.Add(key, value);
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public void Clear() => this.internalDictionary.Clear();
+    public bool ContainsKey(LinkRelation key) => internalDictionary.ContainsKey(key);
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public bool Contains(KeyValuePair<LinkRelation, HttpOperation> item) => this.internalDictionary.ContainsKey(item.Key) && this.internalDictionary[item.Key].Equals(item.Value);
+    public bool Remove(LinkRelation key) => internalDictionary.Remove(key);
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public void CopyTo(KeyValuePair<LinkRelation, HttpOperation>[] array, int arrayIndex) => ((IDictionary)this.internalDictionary).CopyTo(array, arrayIndex);
+    public bool TryGetValue(LinkRelation key, out HttpOperation value) =>
+        internalDictionary.TryGetValue(key, out value);
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        public bool Remove(KeyValuePair<LinkRelation, HttpOperation> item) => this.internalDictionary.Remove(item.Key);
+    public HttpOperation this[LinkRelation key]
+    { get => internalDictionary.ContainsKey(key) ? internalDictionary[key] : null; set => internalDictionary[key] = value; }
 
-        /// <summary>
-        /// <inheritdoc/>
-        /// </summary>
-        IEnumerator<KeyValuePair<LinkRelation, HttpOperation>> IEnumerable<KeyValuePair<LinkRelation, HttpOperation>>.GetEnumerator() => this.internalDictionary.GetEnumerator();
-
-        /// <summary>
-        /// Gets a <paramref name="rel"/> if the dictionary contains it, otherwise null.
-        /// </summary>
-        public HttpOperation this[LinkRelation rel] => ContainsKey(rel) ? this.internalDictionary[rel] : null;
-    }
+    public ICollection<LinkRelation> Keys => internalDictionary.Keys;
+    public ICollection<HttpOperation> Values => internalDictionary.Values;
 }
