@@ -1,13 +1,12 @@
-﻿using System.Reflection;
+using System.Reflection;
 
-namespace SwedbankPay.Sdk
+namespace SwedbankPay.Sdk;
+
+public static class UserAgent
 {
-    public static class UserAgent
+    public static string Default { get; }
+    static UserAgent()
     {
-        public static string Default { get; }
-        static UserAgent()
-        {
-            Default = $"swedbankpay-sdk-dotnet/{Assembly.GetExecutingAssembly().GetCustomAttribute<AssemblyFileVersionAttribute>()?.Version}";
-        }
+        Default = $"swedbankpay-sdk-dotnet/{Assembly.GetExecutingAssembly().GetCustomAttribute<AssemblyFileVersionAttribute>()?.Version}";
     }
 }
