@@ -1,15 +1,12 @@
-﻿using System;
+namespace SwedbankPay.Sdk;
 
-namespace SwedbankPay.Sdk
+/// <summary>
+/// Contains a <seealso cref="Uri"/> to a uniquely identifiable resource.
+/// </summary>
+public interface IIdentifiable
 {
     /// <summary>
-    /// Contains a <seealso cref="Uri"/> to a uniquely identifiable resource.
+    /// The relative URI and unique identifier of the payment resource .
     /// </summary>
-    public interface IIdentifiable
-    {
-        /// <summary>
-        /// The relative URI and unique identifier of the payment resource .
-        /// </summary>
-        Uri Id { get; }
-    }
+    Uri Id { get; }
 }
